extract PushSubscriptionsObserver
