+++ Sources/WalletConnectPush/Client/Wallet/WalletPushClient.swift
@@ -124,22 +124,3 @@ extension WalletPushClient {
 }
 #endif
 
-class PushSubscriptionsObserver {
-    public var subscriptionsPublisher: AnyPublisher<[PushSubscription], Never> {
-        subscriptionsPublisherSubject.eraseToAnyPublisher()
-    }
-    private let subscriptionsPublisherSubject = PassthroughSubject<[PushSubscription], Never>()
-
-    private let store: CodableStore<PushSubscription>
-    
-    init(store: CodableStore<PushSubscription>) {
-        self.store = store
-        setUpSubscription()
-    }
-
-    func setUpSubscription() {
-        store.onStoreUpdate = { [unowned self] in
-            subscriptionsPublisherSubject.send(store.getAll())
-        }
-    }
-}
